+++ src/plugins/piral-oauth2/src/index.ts
@@ -1,3 +1,4 @@
 export * from './create';
 export * from './setup';
 export * from './types';
+export * from './utils';